API 구현 1차 완료

+++ umc9th/src/main/java/com/example/umc9th/domain/review/dto/ReviewReqDTO.java
@@ -9,5 +9,5 @@ public class ReviewReqDTO {
     @Builder
     public record ReviewRequestDTO(
         String content,
-        BigDecimal score){}
+        Float score){}
 }
